+++ sdk/metricsadvisor/Azure.AI.MetricsAdvisor/src/Models/AnomalyDetection/GetAnomaliesForDetectionConfigurationOptions.cs
@@ -2,12 +2,14 @@
 // Licensed under the MIT License.
 
 using System;
+using System.Threading;
 
 namespace Azure.AI.MetricsAdvisor.Models
 {
     /// <summary>
-    /// The set of options that can be specified when calling <see cref="MetricsAdvisorClient.GetAnomaliesForDetectionConfiguration"/>
-    /// or <see cref="MetricsAdvisorClient.GetAnomaliesForDetectionConfigurationAsync"/> to configure the behavior of the request.
+    /// The set of options that can be specified when calling <see cref="MetricsAdvisorClient.GetAnomalies(string, GetAnomaliesForDetectionConfigurationOptions, CancellationToken)"/>
+    /// or <see cref="MetricsAdvisorClient.GetAnomaliesAsync(string, GetAnomaliesForDetectionConfigurationOptions, CancellationToken)"/>
+    /// to configure the behavior of the request.
     /// </summary>
     public class GetAnomaliesForDetectionConfigurationOptions
     {