+++ sdk/metricsadvisor/Azure.AI.MetricsAdvisor/src/Models/AnomalyIncident.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 using System;
+using System.Threading;
 using Azure.Core;
 
 namespace Azure.AI.MetricsAdvisor.Models
@@ -37,8 +38,8 @@ namespace Azure.AI.MetricsAdvisor.Models
         /// <summary>
         /// The unique identifier of the <see cref="AnomalyDetectionConfiguration"/> that detected
         /// this <see cref="AnomalyIncident"/>. This property is only populated when calling
-        /// <see cref="MetricsAdvisorClient.GetIncidentsForAlert"/> or
-        /// <see cref="MetricsAdvisorClient.GetIncidentsForAlertAsync"/>.
+        /// <see cref="MetricsAdvisorClient.GetIncidents(string, string, GetIncidentsForAlertOptions, CancellationToken)"/> or
+        /// <see cref="MetricsAdvisorClient.GetIncidentsAsync(string, string, GetIncidentsForAlertOptions, CancellationToken)"/>.
         /// </summary>
         [CodeGenMember("AnomalyDetectionConfigurationId")]
         public string DetectionConfigurationId { get; }
@@ -46,8 +47,8 @@ namespace Azure.AI.MetricsAdvisor.Models
         /// <summary>
         /// The unique identifier of the <see cref="DataFeedMetric"/> of the time series in which this
         /// <see cref="AnomalyIncident"/> has been detected. This property is only populated when calling
-        /// <see cref="MetricsAdvisorClient.GetIncidentsForAlert"/> or
-        /// <see cref="MetricsAdvisorClient.GetIncidentsForAlertAsync"/>.
+        /// <see cref="MetricsAdvisorClient.GetIncidents(string, string, GetIncidentsForAlertOptions, CancellationToken)"/> or
+        /// <see cref="MetricsAdvisorClient.GetIncidentsAsync(string, string, GetIncidentsForAlertOptions, CancellationToken)"/>.
         /// </summary>
         public string MetricId { get; }